Refactor suggestion classes
