+++ addons/metadata/suggestions/__init__.py
@@ -1,0 +1,2 @@
+from .erad import suggestion_erad
+from .contributor import suggestion_contributor
